PICDRIVE: make save_frames take ULONG num_frames.

+++ src/PJ/inc/picdrive.h
@@ -185,7 +185,7 @@ struct pdr {
 
 	Errcode (*save_frames)(Image_file *ifile,
 						   Rcel *screen,
-						   int num_frames,
+						   ULONG num_frames,
 						   Errcode (*seek_frame)(int ix,void *seek_data),
 						   void *seek_data,
 						   Rcel *work_screen );
@@ -326,7 +326,7 @@ void	pdr_close_ifile(Image_file **pifile);
 Errcode pdr_read_first(Image_file *ifile, Rcel *screen);
 Errcode pdr_read_next(Image_file *ifile,Rcel *screen);
 
-Errcode pdr_save_frames(Image_file *ifile,	Rcel *screen,  int num_frames,
+Errcode pdr_save_frames(Image_file *ifile,	Rcel *screen, ULONG num_frames,
 						Errcode (*seek_frame)(int ix,void *seek_data),
 						void *seek_data, Rcel *work_screen );
