+++ src/PJ/picdrive/pcx/pcx.c
@@ -580,7 +580,7 @@ static Errcode pcx_read_next(Image_file *ifile,Rcel *screen)
 }
 
 
-static Errcode pcx_save_frame(Image_file *ifile, Rcel *screen, int num_frames,
+static Errcode pcx_save_frame(Image_file *ifile, Rcel *screen, ULONG num_frames,
 						      Errcode (*seek_frame)(int ix,void *seek_data),
 						      void *seek_data, Rcel *work_screen ) 
 {
